"Reformat code to use router"
This allows for much more concise code with variables instead of
hard-coding the endpoint for each different subject. The format for the
JSON file was also modified to accommodate for this change. The HTML
file was also slightly changed, but it will be changed even more soon as
it does not have buttons for all the endpoints.